+++ lib/review.js
@@ -1,6 +1,7 @@
 const mongoose = require('mongoose');
 
 const { Reviewer } = require('./role');
+const { releaseForm } = require('./form');
 
 const User = mongoose.model('User');
 const logger = require('./loggers').getLogger();
@@ -46,7 +47,10 @@ async function removeReviewRequest(req, res, doc) {
 async function addReviewResult(req, res, doc) {
   const { result = '2', comment, v } = req.body;
   try {
-    await doc.addReviewResult(req.session.userid, result, comment, v);
+    const newDoc = await doc.addReviewResult(req.session.userid, result, comment, v);
+    if(newDoc.status === 0.5 && newDoc.allApproved) {
+      return await releaseForm(req, res)
+    }
     return res
       .status(201)
       .send(`review result from user ${req.session.userid} added.`);
