+++ routes/form.js
@@ -17,11 +17,11 @@ const reviewLib = require('../lib/review');
 const Form = mongoose.model('Form');
 const FormFile = mongoose.model('FormFile');
 const ReleasedForm = mongoose.model('ReleasedForm');
-const FormContent = mongoose.model('FormContent');
 const User = mongoose.model('User');
 const Group = mongoose.model('Group');
 const History = mongoose.model('History');
 const { stateTransition } = require('../model/form');
+const { releaseForm } = require('../lib/form');
 
 const logger = require('../lib/loggers').getLogger();
 
@@ -1017,107 +1017,7 @@ module.exports = function(app) {
       }
       return next();
     },
-    // if the base form is normal then load the released discrepancy form
-    function(req, res, next) {
-      debug(req.body.discrepancyFormId);
-      debug(req[req.params.id].formType);
-      if (
-        req[req.params.id].formType === 'normal' &&
-        req.body.discrepancyFormId
-      ) {
-        reqUtils.existSource('discrepancyFormId', 'body', ReleasedForm)(
-          req,
-          res,
-          next
-        );
-      } else {
-        next();
-      }
-    },
-    // check the discrepancy form type
-    function(req, res, next) {
-      debug(req[req.body.discrepancyFormId]);
-      if (
-        req[req.body.discrepancyFormId] &&
-        req[req.body.discrepancyFormId].formType !== 'discrepancy'
-      ) {
-        return res
-          .status(400)
-          .send(
-            `${req[req.body.discrepancyFormId].id} is not a discrepancy form`
-          );
-      }
-
-      if (
-        req[req.body.discrepancyFormId] &&
-        req[req.body.discrepancyFormId].status !== 1
-      ) {
-        return res
-          .status(400)
-          .send(`${req[req.body.discrepancyFormId].id} is not released`);
-      }
-      return next();
-    },
-    async function releaseForm(req, res) {
-      const releasedForm = {};
-      const form = req[req.params.id];
-      const discrepancyForm = req[req.body.discrepancyFormId];
-      releasedForm.title = req.body.title || form.title;
-      releasedForm.description = req.body.description || form.description;
-      releasedForm.tags = form.tags;
-      releasedForm.formType = form.formType;
-      releasedForm.base = new FormContent(form);
-      releasedForm.ver = `${releasedForm.base._v}`;
-      if (discrepancyForm) {
-        // update formType
-        releasedForm.formType = 'normal_discrepancy';
-        releasedForm.discrepancy = discrepancyForm.base;
-        releasedForm.ver += `:${discrepancyForm.base._v}`;
-      }
-      releasedForm.releasedBy = req.session.userid;
-      releasedForm.releasedOn = Date.now();
-
-      // check if there is already a released form with the same name and
-      // version
-      try {
-        const existingForm = await ReleasedForm.findOne({
-          title: releasedForm.title,
-          formType: releasedForm.formType,
-          ver: releasedForm.ver,
-          // only search the active released form, not archived
-          // remove this condition if including the archive released form
-          status: 1,
-        });
-        debug(`find existing form: ${existingForm}`);
-        if (existingForm) {
-          return res
-            .status(400)
-            .send(
-              `A form with same title, type, and version was already released in ${existingForm._id}.`
-            );
-        }
-        const saveForm = await new ReleasedForm(releasedForm).saveWithHistory(
-          req.session.userid
-        );
-
-        // update the form status
-        form.status = 1;
-        form.updatedBy = req.session.userid;
-        form.updatedOn = Date.now();
-        await form.save();
-
-        // close the review requests
-        await form.closeReviewRequests();
-        const url = `${
-          req.proxied ? authConfig.proxied_service : authConfig.service
-        }/released-forms/${saveForm._id}/`;
-        return res.status(201).json({
-          location: url,
-        });
-      } catch (error) {
-        return res.status(500).send(error.message);
-      }
-    }
+    releaseForm
   );
 
   app.put(
